Create new formData object
This new object is stored in App and will store
the data from all forms. This way App will be able
to control state, instead of each form having
to do so.

Started creating the logic for controlling state
with the PersonalInfo component. This will be
similar for all forms eventually.