Update App logic
Added a currentEntryIndex state. With this state,
whenever a new entry is added, the state will be
updated with that entry's index. This way, I can
keep track of which entry is being updated in the
handleSectionsUpdate function.

The currentEntryIndex value will be set everytime
a new entry is made, or when a current entry is
selected to edit.

The ExperienceInfo component uses
this index to keep track of the current values of
the entry. This way, it knows which values it
needs to display in the form fields. If it's a new
entry, the form fields will be blank. If it's a
entry that currently exists, the form fields will
use that data as their values.