Update App logic
- Created separate states for displaying each
form. This lead to having to update the logic
for when each form will be displayed.

- Added the editingMode state. Keeps track of when
then user is editing an entry.

- Updated the handleCancel function. If a user is
not in editing mode (otherwise, is adding a new
entry) and presses the cancel button, the blank
entry that was added after clicking the "add"
button will be deleted. This ensures that If
the user and is in editing mode, the entry will
not be deleted.

- The handleCancel function calls on the handle
delete function to remove the blank entry added
when the 'add' button is clicked.